Add Url model
This commit adds the model that will store all the url.
The attributes are:

- target, representing the URL that the user will be redirected
- hits, a count on how many times a URL was used (will be useful to
  create a rank of most used links)
- slug, the record ID represented as kind of Base62 number

About the slug decision:
One of the constraints of the project is that the slug should be the
shortest as possible.

So I took advantage of the database's auto-increment ID to generate the
slug. This way I don't have to worry about uniqueness of the slug and
still sure that it will be the shortest possible.

I also decided to not include some valid url chars (-_.~) just for
convenience. I'm explicity adding the constraint of only using
uppercase and lowercase strings and numbers for now

If we need to add custom slugs in the future, it can be done by a
"custom_slug" column on the database